Fix type warnings
