+++ example/src/navigation/EventsStackParams.tsx
@@ -1,0 +1,4 @@
+export type EventsStackParams = {
+  Events: undefined;
+  EventDetail: undefined;
+};
