+++ example/src/navigation/LoginStackParams.tsx
@@ -1,0 +1,4 @@
+export type LoginStackParams = {
+  Login: undefined;
+  SignUp: undefined;
+};
