+++ example/src/screens/Events.tsx
@@ -1,12 +1,16 @@
 import * as React from 'react';
 import { View, Text, Button } from 'react-native';
+import type { NativeStackScreenProps } from '@react-navigation/native-stack';
 
 import { styles } from '../components/Styles';
+import type { EventsStackParams } from '../navigation/EventsStackParams';
 
-export function EventsScreen({ navigation: { navigate } }) {
+export function EventsScreen({
+  navigation,
+}: NativeStackScreenProps<EventsStackParams, 'Events'>) {
   const handleSelect = () => {
     console.log('Select');
-    navigate('EventDetail');
+    navigation.navigate('EventDetail');
   };
 
   return (
